// resources/js/types/auth.ts
export type User = {
    id: number;
    name: string;
    email: string;
    role: 'admin' | 'staff' | 'user';
    avatar?: string;
    email_verified_at: string | null;
    two_factor_enabled?: boolean;
    date_of_birth: string | null;
    gender: 'male' | 'female' | null;
    created_at: string;
    updated_at: string;
    [key: string]: unknown;
};

export type Auth = {
    user: User;
};

export type TwoFactorSetupData = {
    svg: string;
    url: string;
};

export type TwoFactorSecretKey = {
    secretKey: string;
};